Implemented function to generate Playfair grid.
Generating a Playfair grid works in two phases:

1) Generating a row of characters based on the keyword.
2) Reshaping the row of characters into a 5 x 5 grid.

There are a two assumptions that I made:

1) A Playfair grid will always be 5 x 5. Since one only must consider
unique, uppercase characters, there are only 26 possible values. As a
result, a 5 x 5 grid is viable.
2) Instances of J can be replaced with I. The main source I used is the
Wikipedia article[1] about the cipher. The provided grid follows this
assumption. In the meantime, I will as well.

[1] https://en.wikipedia.org/wiki/Playfair_cipher